Migrate from tokens to cookies
