+++ lib/frontend-app/src/data/user/user-query.impl.ts
@@ -32,22 +32,21 @@ const user_cache: Record<string, Ordo.User.Public.DTO> = {}
 export const UserQuery: Ordo.User.QueryStatic = {
 	Of: check_permission => {
 		const version_zags = ZAGS.Of({ version: 0 })
-		ordo_app_state.zags.cheat("auth.user", (_, is_update) => is_update && version_zags.update("version", i => i + 1))
+		ordo_app_state.zags.cheat("user", (_, is_update) => is_update && version_zags.update("version", i => i + 1))
 
 		const fetch = ordo_app_state.zags.select("fetch")
 		const id_host = ordo_app_state.zags.select("hosts.id")
-		const token = ordo_app_state.zags.select("auth.token")
 
 		return {
 			is_authenticated: () =>
 				check_permission("user.is_authenticated")
-					.pipe(Result.ops.map(() => ordo_app_state.zags.select("auth.user")))
+					.pipe(Result.ops.map(() => ordo_app_state.zags.select("user")))
 					.pipe(Result.ops.map(user => !!user))
 					.cata(Result.catas.or_else(() => false)),
 
 			// TODO Allow user null
 			get_current: () =>
-				check_permission("user.get_current").pipe(Result.ops.chain(() => Result.Ok(ordo_app_state.zags.select("auth.user")))),
+				check_permission("user.get_current").pipe(Result.ops.chain(() => Result.Ok(ordo_app_state.zags.select("user")))),
 
 			get_by_id: id =>
 				check_permission("user.get_by_id")
@@ -60,14 +59,13 @@ export const UserQuery: Ordo.User.QueryStatic = {
 					.and(
 						id =>
 							user_cache[id] ??
-							Oath.Resolve(id)
-								.and(id => ({ id, headers: { Authorization: `Bearer ${token}` } }))
-								.and(({ id, headers }) => Oath.FromPromise(() => fetch(`${id_host}/users/${id}`, { headers })))
+							Oath.FromPromise(() => fetch(`${id_host}/users/${id}`, { credentials: "include" }))
 								.and(res => res.json())
 								.and(res => Oath.If(res.success, { T: () => res.payload, F: () => RRR.codes.eio(res.payload) }))
 								.pipe(ops0.tap(dto => void (user_cache[dto.id] = dto))),
 					)
 					.and(PublicUser.FromDTO),
+
 			get_by_handle: handle =>
 				check_permission("user.get_by_id")
 					.cata({ Ok: () => Oath.Resolve(void 0), Err: rrr => Oath.Reject<Ordo.Rrr<"EPERM">, void>(rrr) })
@@ -79,9 +77,7 @@ export const UserQuery: Ordo.User.QueryStatic = {
 					.and(
 						handle =>
 							user_cache[handle] ??
-							Oath.Resolve(handle)
-								.and(handle => ({ handle, headers: { Authorization: `Bearer ${token}` } }))
-								.and(({ handle, headers }) => Oath.FromPromise(() => fetch(`${id_host}/users/handle/${handle}`, { headers })))
+							Oath.FromPromise(() => fetch(`${id_host}/users/handle/${handle}`, { credentials: "include" }))
 								.and(res => res.json())
 								.and(res => Oath.If(res.success, { T: () => res.payload, F: () => RRR.codes.eio(res.payload) }))
 								.pipe(ops0.tap(dto => void (user_cache[dto.handle] = dto))),
